lab05/requirements.txt
torch==1.10.0
torchvision==0.11.0
pytorch-ignite==0.1.1
tqdm
configargparse